Move float and bool parameters to their own files
